Updated program

+++ src/main/kotlin/com/dannybierek/tools/hmc/model/Position.kt
@@ -1,7 +1,7 @@
 package com.dannybierek.tools.hmc.model
 
 data class Position(
-    val x: Int,
-    val y: Int,
+    val x: Double,
+    val y: Double,
     val z: Double
 )
